# openapi/platform-api.json
{
  "components": {
    "schemas": {
...
      "CreateApplicationRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "workflowId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "workflowId",
          "name",
          "slug",
          "visibility"
        ],
        "type": "object"
      },
...
    },
...
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Agentx Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/applications": {
      "get": {
        "operationId": "list_applications",
        "responses": {},
        "tags": [
          "applications"
        ]
      },
      "post": {
        "operationId": "create_application",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "applications"
        ]
      }
    },
...
  },
...
}